app/requirements.txt
amba-event-stream
pydantic[email]
aiofiles
aiokafka
sqlalchemy